fix(feedback): embed feedback/share in settings + make sharing a forwardable message
用户真机反馈两处与获批样张不符（docs/design/mockups/2026-09-01-feedback-share-center*）：

1) 「反馈与分享」点开变成独立浮层，不在设置页面里。
   根因：AboutSection 的入口 dispatch 全局 nomi-open-feedback-share + onClose()，
   全局 FeedbackShareHost 把 FeedbackShareDialog 渲染成独立 DesignModal。
   修法：把页面路由/表单/分享抽成容器无关的 FeedbackShareContent；About 区块在
   about↔feedback 间切视图、直接内嵌渲染它（不再关设置、不再 dispatch）。样张形态复原：
   反馈长在设置弹窗右栏，左侧 tab 常驻，顶部「‹ 关于」面包屑。
   FeedbackShareHost/Dialog 保留，只服务画布失败卡那条无设置外壳的情境入口（variant='modal'）。

2) 「把 Nomi 发给朋友」给的是裸网站链接。
   修法：分享页新增「一段可直接转发的话」——中文/英文推荐语 + 真实链接，一键复制到剪贴板
   （buildShareMessage 从 i18n 模板 + NOMI_COMMUNITY_LINKS 单一真相源拼出）。zh+en。

placement 断言焊死（本次事故正因走查只验存在不验位置）：feedback-share-center.walk 路径 A
现断言反馈主体是 [data-settings-dialog] 的后代、设置未被关、屏上无另开的反馈 role=dialog，
并验分享文案含真实链接+人话推荐语、点复制后剪贴板即那段话。

- 真机走查通过（zh），截图亲验：内嵌态 + 分享复制态 + success 均在设置容器内
- 全量 gates 通过（9180 unit + contracts + i18n parity + build）
