+++ src/i18n/locales/community.ts
@@ -4,7 +4,7 @@ export const zhCommunity = {
   reportProblem: '告诉我们一件事',
   reportProblemHint: '遇到问题，或有一个具体建议',
   shareNomi: '分享 Nomi',
-  shareNomiHint: '打开官网和 GitHub，发给朋友或团队',
+  shareNomiHint: '一段可直接转发的话，一键复制发给朋友',
   trustLine: '默认只保存在本机；不会后台上传对话、素材或密钥。',
   back: '返回',
   problem: '遇到问题',
@@ -31,7 +31,11 @@ export const zhCommunity = {
   copySummary: '复制摘要',
   required: '先写一句话，方便我们定位',
   shareTitle: '把 Nomi 发给朋友',
-  shareHint: '选择一个入口即可，不需要注册 Nomi 账号。',
+  shareHint: '复制下面这段话，直接发给朋友或团队——不需要注册 Nomi 账号。',
+  // 一段可直接转发的话（问题 #2）：像人话，不是裸链接。{{website}}/{{github}} 由 buildShareMessage 填真实链接。
+  shareMessage: '发你个好东西——Nomi，本地优先的 AI 视频创作工具。一句话就能生成分镜、出片，素材都在自己电脑上，不用等云端排队。\n官网：{{website}}\nGitHub：{{github}}',
+  shareMessageCopy: '复制这段话',
+  shareMessageCopied: '已复制，去粘给朋友吧',
   website: '打开官网',
   github: '查看 GitHub',
   localDraft: '没有后台上传；你选择的目标页面会在浏览器中单独确认。',
@@ -43,7 +47,7 @@ export const enCommunity = {
   reportProblem: 'Tell us one thing',
   reportProblemHint: 'Report a problem or make a concrete suggestion',
   shareNomi: 'Share Nomi',
-  shareNomiHint: 'Open the website and GitHub for a friend or teammate',
+  shareNomiHint: 'A ready-to-send message you can copy in one tap',
   trustLine: 'Saved locally by default; conversations, assets, and keys are never uploaded in the background.',
   back: 'Back',
   problem: 'Something broke',
@@ -70,7 +74,11 @@ export const enCommunity = {
   copySummary: 'Copy summary',
   required: 'Add one sentence so we can find the right path',
   shareTitle: 'Share Nomi with a friend',
-  shareHint: 'Pick an entry point; no Nomi account is required.',
+  shareHint: 'Copy the message below and send it straight to a friend or teammate — no Nomi account required.',
+  // A paste-ready message (issue #2): reads like a person, not a bare link. buildShareMessage fills the real URLs.
+  shareMessage: 'Found something you should try — Nomi, a local-first AI video studio. Describe a shot in a sentence and it storyboards and renders it, with your footage staying on your own machine, no cloud queue.\nSite: {{website}}\nGitHub: {{github}}',
+  shareMessageCopy: 'Copy this message',
+  shareMessageCopied: 'Copied — go paste it to a friend',
   website: 'Open website',
   github: 'View GitHub',
   localDraft: 'No background upload; the destination page is confirmed separately in your browser.',
